Updated all extensions to latest SDK changes

+++ source/Libraries/EpicLibrary/Services/EpicAccountClient.cs
@@ -23,12 +23,18 @@ namespace EpicLibrary.Services
         }
     }
 
+    public class ApiRedirectResponse
+    {
+        public string redirectUrl { get; set; }
+        public string sid { get; set; }
+    }
+
     public class EpicAccountClient
     {
         private ILogger logger = LogManager.GetLogger();
         private IPlayniteAPI api;
         private string tokensPath;
-        private readonly string loginUrl = @"";
+        private readonly string loginUrl = "https://www.epicgames.com/id/login?redirectUrl=https://www.epicgames.com/id/api/redirect";
         private readonly string oauthUrl = @"";
         private readonly string accountUrl = @"";
         private readonly string assetsUrl = @"";
@@ -39,7 +45,6 @@ namespace EpicLibrary.Services
         {
             this.api = api;
             this.tokensPath = tokensPath;
-            var loginUrlMask = @"https://{0}/login?redirectUrl=https%3A%2F%2F{0}%2Flogin%2FshowPleaseWait%3Fclient_id%3D24a1bff3f90749efbfcbc576c626a282%26rememberEmail%3Dfalse&client_id=24a1bff3f90749efbfcbc576c626a282&isLauncher=true";
             var oauthUrlMask = @"https://{0}/account/api/oauth/token";
             var accountUrlMask = @"https://{0}/account/api/public/account/";
             var assetsUrlMask = @"https://{0}/launcher/api/public/assets/Windows?label=Live";
@@ -70,22 +75,20 @@ namespace EpicLibrary.Services
                 assetsUrl = string.Format(assetsUrlMask, "launcher-public-service-prod06.ol.epicgames.com");
                 catalogUrl = string.Format(catalogUrlMask, "catalog-public-service-prod06.ol.epicgames.com");
             }
-
-            loginUrl = string.Format(loginUrlMask, "accounts.launcher-website-prod07.ol.epicgames.com");
         }
 
         public void Login()
         {
             var loggedIn = false;
-            var loginPageSource = string.Empty;
+            var apiRedirectContent = string.Empty;
             using (var view = api.WebViews.CreateView(580, 700))
             {
                 view.LoadingChanged += async (s, e) =>
                 {
                     var address = view.GetCurrentAddress();
-                    if (address.StartsWith(@"https://accounts.launcher-website-prod07.ol.epicgames.com/login/showPleaseWait"))
+                    if (address.StartsWith(@"https://www.epicgames.com/id/api/redirect"))
                     {
-                        loginPageSource = await view.GetPageSourceAsync();
+                        apiRedirectContent = await view.GetPageTextAsync();
                         loggedIn = true;
                         view.Close();
                     }
@@ -101,8 +104,14 @@ namespace EpicLibrary.Services
                 return;
             }
 
+            if (apiRedirectContent.IsNullOrEmpty())
+            {
+                return;
+            }
+
+            var sid = Serialization.FromString<ApiRedirectResponse>(apiRedirectContent, Format.Json).sid;
             FileSystem.DeleteFile(tokensPath);
-            var exchangeKey = getExcahngeToken(loginPageSource);
+            var exchangeKey = getExcahngeToken(sid);
             if (string.IsNullOrEmpty(exchangeKey))
             {
                 logger.Error("Failed to get login exchange key for Epic account.");
@@ -172,7 +181,7 @@ namespace EpicLibrary.Services
             {
                 try
                 {
-                    result = Serialization.FromJsonFile<Dictionary<string, CatalogItem>>(cachePath);
+                    result = Serialization.FromFile<Dictionary<string, CatalogItem>>(cachePath, Format.Json);
                 }
                 catch (Exception e)
                 {
@@ -221,7 +230,7 @@ namespace EpicLibrary.Services
             }
         }
 
-        private async Task<Tuple<string, T>> InvokeRequest<T>(string url, OauthResponse tokens) where T : class, new()
+        private async Task<Tuple<string, T>> InvokeRequest<T>(string url, OauthResponse tokens) where T : class
         {
             using (var httpClient = new HttpClient())
             {
@@ -230,22 +239,13 @@ namespace EpicLibrary.Services
                 var response = await httpClient.GetAsync(url);
                 var str = await response.Content.ReadAsStringAsync();
 
-                ErrorResponse error = null;
-                try
-                {
-                    error = Serialization.FromJson<ErrorResponse>(str);
-                }
-                catch
-                {
-                }
-
-                if (error != null && !string.IsNullOrEmpty(error.errorCode))
+                if (Serialization.TryFromString<ErrorResponse>(str, Format.Json, out var error) && !string.IsNullOrEmpty(error.errorCode))
                 {
                     throw new TokenException(error.errorCode);
                 }
                 else
                 {
-                    return new Tuple<string, T>(str, Serialization.FromJson<T>(str));
+                    return new Tuple<string, T>(str, Serialization.FromString<T>(str, Format.Json));
                 }
             }
         }
@@ -257,18 +257,38 @@ namespace EpicLibrary.Services
                 return null;
             }
 
-            return Serialization.FromJson<OauthResponse>(File.ReadAllText(tokensPath));
+            return Serialization.FromFile<OauthResponse>(tokensPath, Format.Json);
         }
 
-        private string getExcahngeToken(string input)
+        private string getExcahngeToken(string sid)
         {
-            var match = Regex.Match(input, @"loginWithExchangeCode\(\'(.*?)\'");
-            if (match.Success)
+            using (var handler = new HttpClientHandler())
+            using (var httpClient = new HttpClient())
             {
-                return match.Groups[1].Value;
-            }
+                httpClient.DefaultRequestHeaders.Clear();
+                httpClient.DefaultRequestHeaders.Add("X-Epic-Event-Action", "login");
+                httpClient.DefaultRequestHeaders.Add("X-Epic-Event-Category", "login");
+                httpClient.DefaultRequestHeaders.Add("X-Epic-Strategy-Flags", "");
+                httpClient.DefaultRequestHeaders.Add("X-Requested-With", "XMLHttpRequest");
+                httpClient.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:86.0) Gecko/20100101 Firefox/86.0");
 
-            return string.Empty;
+                httpClient.GetAsync(@"https://www.epicgames.com/id/api/set-sid?sid=" + sid).GetAwaiter().GetResult();
+                var resp = httpClient.GetAsync(@"https://www.epicgames.com/id/api/csrf").GetAwaiter().GetResult();
+                var cookies = resp.Headers.Single(header => header.Key == "Set-Cookie").Value;
+                if (cookies != null)
+                {
+                    var match = Regex.Match(cookies.First(), @"=(.+);");
+                    var xsrf = match.Groups[1].Value;
+                    httpClient.DefaultRequestHeaders.Add("X-XSRF-TOKEN", xsrf);
+                    resp = httpClient.PostAsync("https://www.epicgames.com/id/api/exchange/generate", null).GetAwaiter().GetResult();
+                    var respContent = resp.Content.ReadAsStringAsync().GetAwaiter().GetResult();
+                    return Serialization.FromString<Dictionary<string, string>>(respContent, Format.Json)["code"];
+                }
+                else
+                {
+                    return null;
+                }
+            }
         }
     }
 }
